+++ nazha/nazha-async/pom.xml
@@ -9,8 +9,8 @@
 		<artifactId>nazha</artifactId>
 		<version>0.0.1-SNAPSHOT</version>
 	</parent>
-	<artifactId>nazha-asyn</artifactId>
-	<name>nazha-asyn</name>
+	<artifactId>nazha-async</artifactId>
+	<name>nazha-async</name>
 	<url>http://maven.apache.org</url>
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
